pyproject.toml
[tool.poetry]
name = "co-dev-backend"
version = "0.1.0"
description = "The backend for co-dev teammates search tool"
authors = []

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.91.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
